requirements.txt
# Core Web Framework
flask
flask-cors
python-dotenv
werkzeug

# Document Processing
python-docx
PyPDF2

# HTTP Requests for LLM APIs
requests

# Data Processing
pyyaml

# Development and Testing
pytest
